+++ src/ai/backend/common/utils.py
@@ -9,7 +9,6 @@ import re
 import sys
 import uuid
 from collections import OrderedDict
-from dataclasses import fields, is_dataclass
 from datetime import timedelta
 from itertools import chain
 from pathlib import Path
@@ -17,23 +16,18 @@ from typing import (
     TYPE_CHECKING,
     Any,
     AsyncIterator,
-    Callable,
     Iterable,
     Iterator,
     Mapping,
-    MutableMapping,
     Optional,
     Tuple,
-    Type,
     TypeVar,
     Union,
-    cast,
 )
 
 import aiofiles
 import yarl
 from async_timeout import timeout
-from graphql import Undefined
 
 if TYPE_CHECKING:
     from decimal import Decimal
@@ -440,54 +434,3 @@ def b64encode(s: str) -> str:
     """
     b: bytes = s.encode("utf-8") if isinstance(s, str) else s
     return base64.b64encode(b).decode("ascii")
-
-
-T = TypeVar("T")
-
-
-class Unset:
-    def __repr__(self):
-        return "UNSET"
-
-
-UNSET = Unset()
-
-
-def graphene_input_to_dataclass(cls: Type[T], data: MutableMapping[str, Any]) -> T:
-    """ """
-    if not is_dataclass(cls):
-        raise ValueError(f"{cls} is not a dataclass type.")
-
-    result: dict[str, Any] = {}
-    for field in fields(cls):
-        value = data.get(field.name, Undefined)
-        if value is Undefined:
-            result[field.name] = None
-        elif value is None:
-            result[field.name] = UNSET
-        else:
-            result[field.name] = value
-
-    return cast(T, cls(**result))
-
-
-def apply_dataclass_field(
-    src: Any,
-    target: dict[str, Any],
-    name: str,
-    *,
-    clean_func: Optional[Callable[[Any], Any]] = None,
-    target_key: Optional[str] = None,
-) -> None:
-    if not (is_dataclass(src) and not isinstance(src, type)):
-        raise TypeError("src must be a dataclass instance.")
-
-    # "name" should exist in the "src"
-    v = getattr(src, name)
-
-    if v is None:
-        pass
-    elif v is UNSET:
-        target[target_key or name] = None
-    else:
-        target[target_key or name] = clean_func(v) if callable(clean_func) else v
