+++ src/ai/backend/manager/services/image/actions/modify_image.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
 from typing import Optional, override
 
-from ai.backend.common.utils import Unset
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.actions.exceptions import BaseActionException
 from ai.backend.manager.data.image.types import ImageData
