+++ tests/manager/services/test_image.py
@@ -5,9 +5,9 @@ from dataclasses import replace
 import pytest
 from dateutil.parser import isoparse
 
-from ai.backend.common.utils import UNSET
 from ai.backend.manager.api.exceptions import ImageNotFound
 from ai.backend.manager.data.image.types import ImageAliasData, ImageData
+from ai.backend.manager.models.base import UNSET
 from ai.backend.manager.models.image import ImageAliasRow, ImageRow, ImageStatus, ImageType
 from ai.backend.manager.models.user import UserRole
 from ai.backend.manager.server import (
